fix: KNNAreaMaskBuilder working with NPZ
